+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 CHANGELOG
 =========
 
-4.1.0
+4.1.0 (2026-01-20)
 ------------------
 
+* Added `anonymizer` property to `IpAddress` response model. This contains
+  information about whether the IP address is an anonymous network, including
+  confidence score, VPN provider name, and various anonymizer flags.
 * Added `BANQUEST`, `SUMMIT_PAYMENTS`, and `YAADPAY` to the
   `Payment.Processor` enum.